% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar.R
\name{check_count}
\alias{check_count}
\title{Checks Count}
\usage{
check_count(x, x_name = substitute(x))
}
\arguments{
\item{x}{The count to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Count
}
\seealso{
\code{\link{datacheckr}}
}
