% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector.R
\name{check_counts}
\alias{check_counts}
\title{Checks Counts}
\usage{
check_counts(x, x_name = substitute(x))
}
\arguments{
\item{x}{The count(s) to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Counts
}
\examples{
check_counts(1L)
}
