% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{check_data}
\alias{check_data}
\title{Check Data}
\usage{
check_data(data, values = NULL, min_row = 0, max_row = max_nrow(),
  key = character(0), data_name = substitute(data))
}
\arguments{
\item{data}{The data frame to check.}

\item{values}{A named list specifying the columns and
their associated classes and values.}

\item{min_row}{A count of the minimum number of rows.}

\item{max_row}{A count of the maximum number of rows.}

\item{key}{A character vector of the column names representing the key.}

\item{data_name}{A string of the name of data.}
}
\value{
Throws an informative error or returns an invisible copy of
data.
}
\description{
Checks a data frame's rows, columns, classes and values and key.
}
\seealso{
\code{\link{datacheckr}}
}

