% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data3.R
\name{check_data3}
\alias{check_data3}
\title{Check Data3}
\usage{
check_data3(data, values = NULL, min_row = 1, max_row = max_nrow(),
  key = character(0), select = FALSE, data_name = substitute(data))
}
\arguments{
\item{data}{The data frame to check.}

\item{values}{A named list specifying the columns and
their associated classes and values.}

\item{min_row}{A count of the minimum number of rows.}

\item{max_row}{A count of the maximum number of rows.}

\item{key}{A character vector of the column names representing the key.}

\item{select}{deprecated A flag indicating whether to if possible
drop unnamed columns and reorder the remainder
so that column names matches those in values.}

\item{data_name}{A string of the name of data.}
}
\value{
Throws an informative error or returns an invisible copy of
data.
}
\description{
Checks a data frame's rows, columns, classes and values and key.
}
\details{
\code{check_data3} enforces maximal constraints.
By default it requires data to have rows (and therefore columns).
It also does not allow NULLs in values and
requires column names in values to be unique.
In addition it also requires that data only includes the columns in values in
the same order.

For more permissive data checking see \code{\link{check_data1}} and
\code{\link{check_data2}}.
}
