% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data-frame.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Check Data Frame}
\usage{
check_data_frame(data, data_name = substitute(data))
}
\arguments{
\item{data}{The data frame to check.}

\item{data_name}{A string of the name of data.}
}
\value{
Throws an informative error or returns an invisible copy of
data.
}
\description{
Checks that data is a data frame.
}
