% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar.R
\name{check_date}
\alias{check_date}
\title{Checks Date}
\usage{
check_date(x, x_name = substitute(x))
}
\arguments{
\item{x}{The Date to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Date
}
\seealso{
\code{\link{datacheckr}}
}

