% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector.R
\name{check_dates}
\alias{check_dates}
\title{Checks Dates}
\usage{
check_dates(x, x_name = substitute(x))
}
\arguments{
\item{x}{The Date(s) to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Dates
}
