% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar.R
\name{check_flag}
\alias{check_flag}
\title{Checks Flag}
\usage{
check_flag(x, x_name = substitute(x))
}
\arguments{
\item{x}{The flag to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Flag
}
\seealso{
\code{\link{datacheckr}}
}

