% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector.R
\name{check_flags}
\alias{check_flags}
\title{Checks Flags}
\usage{
check_flags(x, x_name = substitute(x))
}
\arguments{
\item{x}{The flag(s) to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Flags
}
\examples{
check_flags(TRUE)
}
