% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.r
\name{check_foreign_keys}
\alias{check_foreign_keys}
\title{Check foreign key values are in reference columns}
\usage{
check_foreign_keys(dt, ref, keys, ref_keys = keys, optional = FALSE)
}
\arguments{
\item{dt}{a data.table, for which to check foreign key values.}

\item{ref}{a data.table, which includes the reference columns to check
against.}

\item{keys}{a character vector, giving names of foreign key columns in
\code{dt}.}

\item{ref_keys}{a character vector, giving names of reference columns in
\code{ref}. These should be in the same order as the foreign key columns
they are references for. Defaults to \code{keys}.}

\item{optional}{a logical, indicating whether foreign key values can be
missing. Note that this is different from the reference columns being
nullable. Defaults to FALSE.}
}
\value{
NULL, if no reference errors found.
}
\description{
Check foreign key values are in reference columns
}
