% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar.R
\name{check_int}
\alias{check_int}
\title{Checks Int}
\usage{
check_int(x, x_name = substitute(x))
}
\arguments{
\item{x}{The scalar integer to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Int
}
\examples{
try(check_int(1))
}
\seealso{
\code{\link{datacheckr}}
}
