% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector.R
\name{check_ints}
\alias{check_ints}
\title{Checks Ints}
\usage{
check_ints(x, x_name = substitute(x))
}
\arguments{
\item{x}{The integer(s) to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Ints
}
\examples{
try(check_ints(1))
}
