% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-key.R
\name{check_key}
\alias{check_key}
\title{Check Key}
\usage{
check_key(data, values = NULL)
}
\arguments{
\item{data}{The data frame to check.}

\item{values}{A named list specifying the columns of the key and
their associated classes and values.}
}
\value{
Throws an informative error or returns an invisible copy of
the original data frame.
}
\description{
Checks that the columns in a data frame represent a unique key.
}
\seealso{
\code{\link{datacheckr}}
}

