% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-length.R
\name{check_length}
\alias{check_length}
\title{Check Length}
\usage{
check_length(x, min_length = 1L, max_length = max_integer(),
  x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{min_length}{A count of the minimum length.}

\item{max_length}{A count of the maximum length.}

\item{x_name}{A string of the name of the object.}
}
\value{
Throws an informative error or returns an invisible copy of
the object.
}
\description{
Checks length of an object is within bounds.
}
\examples{
try(check_length(1:2, 3L))
}
