% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-null.R
\name{check_null}
\alias{check_null}
\title{Check NULL}
\usage{
check_null(x, x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}
}
\value{
Throws an informative error or returns NULL.
}
\description{
Checks object is NULL.
}
