% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar.R
\name{check_number}
\alias{check_number}
\title{Checks Number}
\usage{
check_number(x, x_name = substitute(x))
}
\arguments{
\item{x}{The number to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Number
}
\seealso{
\code{\link{datacheckr}}
}

