% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector.R
\name{check_numbers}
\alias{check_numbers}
\title{Checks Number}
\usage{
check_numbers(x, x_name = substitute(x))
}
\arguments{
\item{x}{The number(s) to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Number
}
\examples{
check_numbers(2)
}
