% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.r
\name{check_primary_keys_unique}
\alias{check_primary_keys_unique}
\title{Check primary keys are unique}
\usage{
check_primary_keys_unique(dt, colnames)
}
\arguments{
\item{dt}{a data.table, for which to check the primary key values.}

\item{colnames}{a character vector, giving names of columns in the primary
key.}
}
\value{
NULL, if no duplicates found.
}
\description{
Check primary keys are unique
}
