% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.r
\name{check_range_contiguous}
\alias{check_range_contiguous}
\title{Check numeric/date start/end interval columns make contiguous interval}
\usage{
check_range_contiguous(dt, start_column, end_column, spacing, by = NULL)
}
\arguments{
\item{dt}{a data.table, from which to check the range columns.}

\item{start_column}{a character, giving the name of the interval start point
column.}

\item{end_column}{a character, giving the name of the interval end point
column.}

\item{spacing}{a numeric or integer vector, giving the expected gap between
an interval end and the next interval's start.}

\item{by}{a character vector, giving names of columns to group over. Defaults
to NULL, for no grouping.}
}
\value{
NULL, if the intervals are contiguous.
}
\description{
Check numeric/date start/end interval columns make contiguous interval
}
