% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar.R
\name{check_scalar}
\alias{check_scalar}
\title{Check Scalar}
\usage{
check_scalar(scalar, value, scalar_name = substitute(scalar))
}
\arguments{
\item{scalar}{The scalar to check.}

\item{value}{A vector specifying the class and value.}

\item{scalar_name}{A string of the name of vector.}
}
\value{
Throws an informative error or returns an invisible copy of
the scalar.
}
\description{
Checks the class and values of a scalar
}
\seealso{
\code{\link{datacheckr}}
}
