% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar.R
\name{check_string}
\alias{check_string}
\title{Checks String}
\usage{
check_string(x, x_name = substitute(x))
}
\arguments{
\item{x}{The string to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks String
}
\seealso{
\code{\link{datacheckr}}
}
