% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector.R
\name{check_strings}
\alias{check_strings}
\title{Checks Strings}
\usage{
check_strings(x, x_name = substitute(x))
}
\arguments{
\item{x}{The string(s) to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Strings
}
\examples{
check_strings("oeu")
}
