% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.r
\name{check_table_constraint}
\alias{check_table_constraint}
\title{Check table entries satisfy given constraint}
\usage{
check_table_constraint(dt, expr)
}
\arguments{
\item{dt}{a data.table, on which to check the constraint.}

\item{expr}{an expression, which will be evaluated to check the constraint.
Columns in the table can be referred to by name.}
}
\value{
NULL, if the constraint is always satisfied.
}
\description{
Check table entries satisfy given constraint
}
