% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar.R
\name{check_time}
\alias{check_time}
\title{Checks Time}
\usage{
check_time(x, x_name = substitute(x))
}
\arguments{
\item{x}{The scalar POSIXct to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Time
}
\examples{
try(check_time(Sys.Date()))
}
\seealso{
\code{\link{datacheckr}}
}
