% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector.R
\name{check_times}
\alias{check_times}
\title{Checks Times}
\usage{
check_times(x, x_name = substitute(x))
}
\arguments{
\item{x}{The scalar POSIXct to check.}

\item{x_name}{A string of the name of x.}
}
\value{
Throws an informative error or returns an invisible copy of
x.
}
\description{
Checks Times
}
\examples{
try(check_times(Sys.Date()))
}
