% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-unique.R
\name{check_unique}
\alias{check_unique}
\title{Check Unique}
\usage{
check_unique(x, x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}
}
\value{
Throws an informative error or returns an invisible copy of
the object.
}
\description{
Checks whether an object in unique (i.e. doesn't contain any duplicated elements).
}
\seealso{
\code{\link{datacheckr}}
}
