% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-values.R
\name{check_values}
\alias{check_values}
\title{Check Values}
\usage{
check_values(data, values = NULL, unique = TRUE, nulls = FALSE,
  data_name = substitute(data))
}
\arguments{
\item{data}{The data frame to check.}

\item{values}{A named list specifying the columns and
their associated classes and values.}

\item{unique}{A string indicating if columns must only be defined in values.}

\item{nulls}{A string indicating if values can be NULL.}

\item{data_name}{A string of the name of data.}
}
\value{
Throws an informative error or returns an invisible copy of
the data.
}
\description{
Checks the class and values of columns in data.
}
\seealso{
\code{\link{datacheckr}}
}

