% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector.R
\name{check_vector}
\alias{check_vector}
\title{Check Vector}
\usage{
check_vector(vector, value, vector_name = substitute(vector))
}
\arguments{
\item{vector}{The vector to check.}

\item{value}{A vector specifying the class and values.}

\item{data_name}{A string of the name of vector.}
}
\value{
Throws an informative error or returns an invisible copy of
the vector.
}
\description{
Checks the class and values of a vector
}
\seealso{
\code{\link{datacheckr}}
}

