% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkor.R
\name{checkor}
\alias{checkor}
\title{Multiple Checks}
\usage{
checkor(...)
}
\arguments{
\item{...}{Checks to check.}
}
\value{
Throws an informative error or returns an invisible TRUE.
}
\description{
Checks that one or more checks return without an error.
Otherwise throws an error consisting of the error messages separated by OR.
}
