% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoTableChecks.R
\name{compareDimensions}
\alias{compareDimensions}
\title{compareDimensions}
\usage{
compareDimensions(table1, table2, commonColumns, columnsToTest,
  columnAggregations, tolerance)
}
\arguments{
\item{table1}{First table to compare}

\item{table2}{Second table to compare}

\item{commonColumns}{A string or vector giving dimensions which should be used to compare both tables}

\item{columnsToTest}{A string or vector giving the columns which should be tested for each common set of dimensions.}

\item{columnAggregations}{A list of aggregations and corresponding labels to be performed on each in column given in columnsToTest as the dataset is rolled up}

\item{testTolerance}{A single value  or vector giving the tolerance to be used to assess whether a test 'passes' or not. If a single value is given this will be used for all columns to be tested.}
}
\description{
compareDimensions
}
\seealso{
compareRows Does a row by row comparison with no pre-aggregation of the data
}
