% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowLevelChecks.R
\name{compareRows}
\alias{compareRows}
\title{Function to do matched row based comparison between files}
\usage{
compareRows(table1, table2, commonKey, testColumn, testType, testTolerance)
}
\arguments{
\item{commonKey}{A string or vector which specify the field(s) which should be used to join rows in between the two tables. The key must be unique in both tables to avoid problems.}

\item{testType}{A string or vector specifying the type of test to be performmed on each field. Currently supports "numeric"}

\item{testTolerance}{A string or vector specifying the tolerance that should be applied for each tested column}
}
\description{
Function to do matched row based comparison between files
}
