% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deck}
\alias{deck}
\title{Data Check}
\usage{
deck(data, values = NULL)
}
\arguments{
\item{data}{The data frame to check.}

\item{values}{A named list specifying the columns and their associated values.}
}
\value{
Throws an informative error or returns the original data object.
}
\description{
Data Check
}

