% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_integer}
\alias{max_integer}
\title{Maximum Integer Value}
\usage{
max_integer()
}
\value{
A count of the maximum integer value.
}
\description{
A wrapper for \code{.Machine$integer.max}
which defines the maximum integer value for the machine.
}
\examples{
max_integer()
}
