% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_integer}
\alias{min_integer}
\title{Minimum Integer Value}
\usage{
min_integer()
}
\value{
A int of the minimum integer value.
}
\description{
A wrapper for \code{-.Machine$integer.max}
which defines the minimum integer value for the machine.
}
\examples{
min_integer()
}
\seealso{
\code{\link{datacheckr}}
}
