% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationcheckr-package.r
\docType{package}
\name{relationchecker-package}
\alias{relationchecker}
\alias{relationchecker-package}
\title{Check relation table data for errors}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
