% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationcheckr-package.r
\docType{package}
\name{relationcheckr-package}
\alias{relationcheckr}
\alias{relationcheckr-package}
\title{Check relation table data for errors}
\description{
A package for simple consistency checks on relational data,
    including checking for missing entries in non-nullable columns and
    errors in foreign keys.
}
\keyword{internal}
