% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cols.R
\name{check_cols}
\alias{check_cols}
\title{Check Cols}
\usage{
check_cols(data, colnames = NULL, exclusive = FALSE, ordered = FALSE,
  data_name = substitute(data))
}
\arguments{
\item{data}{The data frame to check.}

\item{colnames}{A character vector of the column names.}

\item{data_name}{A string of the name of data.}
}
\value{
Throws an informative error or returns an invisible copy of
the data.
}
\description{
Checks the names and order of columns in data.
}
\details{
By default (\code{colnames = NULL}) data must include at least one column.
To check for no columns set \code{colnames = character(0)}.
}
\seealso{
\code{\link{datacheckr}}
}

