% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{check_data}
\alias{check_data}
\title{Check Data}
\usage{
check_data(data, values = NULL)
}
\arguments{
\item{data}{The data frame to check.}

\item{values}{A named list specifying the columns and their associated values.}
}
\value{
Throws an informative error or returns the original data object.
}
\description{
Checks data based on a named
list of vectors that specifies the possible columns and their associated classes
and values.
If data passes all the conditions set by values then data_check returns the original
data frame (which allows the function to be inserted in piping chains).
Otherwise data_check throws an informative error.
}
\details{
If values is unspecified data_check simply checks that data is a data frame.
The regular expression checking of character strings is performed using
grepl with perl = TRUE.
}

