% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data2.R
\name{check_data2}
\alias{check_data2}
\title{Check Data2}
\usage{
check_data2(data, values = NULL, min_row = 0, max_row = max_nrow(),
  key = NULL, data_name = substitute(data))
}
\arguments{
\item{data}{The data frame to check.}

\item{key}{A character vector of the column names representing the key.}

\item{data_name}{A string of the name of data.}
}
\value{
Throws an informative error or returns an invisible copy of
data.
}
\description{
Checks a data frame's rows columns, classes, values, and key.
}
\seealso{
\code{\link{datacheckr}}
}

