% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-join.R
\name{check_join}
\alias{check_join}
\title{Check Join}
\usage{
check_join(data, parent, key = NULL, referential = TRUE, extra = FALSE,
  data_name = substitute(data), parent_name = substitute(parent))
}
\arguments{
\item{data}{The data frame to check.}

\item{parent}{A data frame of the parent table.}

\item{key}{A character vector of the key columns.}

\item{referential}{A flag indicating whether to check for referential integrity.}

\item{extra}{A flag indicating whether to allow additional matching columns.}

\item{data_name}{A string of the name of data.}

\item{parent_name}{A string of the name of parent.}
}
\value{
Throws an informative error or returns an invisible copy of
data.
}
\description{
Checks that the columns in a data frame form a many-to-one
join with correponding columns in parent.
}
\details{
By default (\code{key = NULL}) all the columns in parent represent the key.
}
\seealso{
\code{\link{datacheckr}}
}

