% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-join.R
\name{check_join}
\alias{check_join}
\title{Check Joins}
\usage{
check_join(data, key, colnames = NULL, data_name = substitute(data),
  key_name = substitute(key))
}
\arguments{
\item{data}{The data frame to check.}

\item{key}{A data.frame representing the key.}

\item{colnames}{A character vector specifying the columns of the key.}

\item{data_name}{A string of the name of data.}

\item{key_name}{A string of the name of key.}
}
\value{
Throws an informative error or returns an invisible copy of
the original data frame.
}
\description{
Checks that the columns in a data frame form a many-to-one
join with correponding columns in the key.
}
\seealso{
\code{\link{datacheckr}}
}

