% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-key.R
\name{check_key}
\alias{check_key}
\title{Check Key}
\usage{
check_key(data, colnames = NULL, data_name = substitute(data))
}
\arguments{
\item{data}{The data frame to check.}

\item{colnames}{A character vector specifying the columns of the key.}
}
\value{
Throws an informative error or returns an invisible copy of
the original data frame.
}
\description{
Checks that the columns in a data frame represent a unique key.
To check the classes and values of the key columns use \code{\link{check_data}}.
}
\seealso{
\code{\link{datacheckr}}
}

