% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-key.R
\name{check_key}
\alias{check_key}
\title{Check Key}
\usage{
check_key(data, key = NULL, data_name = substitute(data))
}
\arguments{
\item{data}{The data frame to check.}

\item{key}{A character vector of the column names representing the key.}

\item{data_name}{A string of the name of data.}
}
\value{
Throws an informative error or returns an invisible copy of
the data.
}
\description{
Checks that columns in a data frame represent a unique key.
}
\details{
By default (\code{key = NULL}) all the columns are checked.
}
\seealso{
\code{\link{datacheckr}}
}
