% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoTableChecks.R
\name{compareRows}
\alias{compareRows}
\title{CompareRows}
\usage{
compareRows(table1, table2, commonKey, testColumn, testType, testTolerance)
}
\arguments{
\item{commonKey}{A string or vector which specify the field(s) which should be used to join rows in between the two tables. The key must be unique in both tables to avoid problems.}

\item{testType}{A string or vector specifying the type of test to be performmed on each field. Currently supports "numeric"}

\item{testTolerance}{A string or vector specifying the tolerance that should be applied for each tested column}
}
\description{
Function to do matched row based comparison between files without any preceding transformation to the data to be compared. This is in contrast to the function compareDimensions which first aggegates both tables over a common set of dimensions.
}
\seealso{
compareDimensions
}
